+++ boa_engine/src/vm/opcode/meta/mod.rs
@@ -16,7 +16,7 @@ pub(crate) struct NewTarget;
 impl Operation for NewTarget {
     const NAME: &'static str = "NewTarget";
     const INSTRUCTION: &'static str = "INST - NewTarget";
-    const COST: usize = 2;
+    const COST: u8 = 2;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         let new_target = if let Some(new_target) = context
@@ -45,7 +45,7 @@ pub(crate) struct ImportMeta;
 impl Operation for ImportMeta {
     const NAME: &'static str = "ImportMeta";
     const INSTRUCTION: &'static str = "INST - ImportMeta";
-    const COST: usize = 6;
+    const COST: u8 = 6;
 
     fn execute(context: &mut Context<'_>) -> JsResult<CompletionType> {
         // Meta Properties
